requirements.txt
numpy
pandas
opencv-python
tensorflow
keras
scikit-learn
matplotlib
notebook
tkinter
pillow
